Move doc building utilities to new file
